[blas] separate helper functions into .hpp
